Time: 67 ms (58.42%), Space: 195.5 MB (58.68%) - LeetHub

+++ 3623-count-number-of-trapezoids-i/3623-count-number-of-trapezoids-i.cpp
@@ -28,13 +28,13 @@ public:
             mpp[point[1]]++;
         }
         long long sum=0;
-        int ans=0;
+        long long ans=0;
         for(auto it=mpp.begin();it!=mpp.end();it++){
            if(it->second>=2){
                ans=(ans+(sum%mod*(((it->second)*(it->second-1))/2)%mod)%mod)%mod;
                sum=(sum%mod+(((it->second)*(it->second-1))/2)%mod)%mod;
            }
         }
-        return ans;
+        return (int)ans;
     }
 };
